restructure jobs into pipelines (multiple jobs achieve one thing) and tasks (single async tasks)
