+++ jobs/pipelines/analyse_lecture/transcribe_audio.py
