+++ jobs/tasks/lecture/clean_lecture.py
